+++ src/main/java/net/geforcemods/securitycraft/SCContent.java
@@ -201,6 +201,7 @@ import net.geforcemods.securitycraft.items.UniversalBlockReinforcerItem;
 import net.geforcemods.securitycraft.items.UniversalBlockRemoverItem;
 import net.geforcemods.securitycraft.items.UniversalKeyChangerItem;
 import net.geforcemods.securitycraft.items.UniversalOwnerChangerItem;
+import net.geforcemods.securitycraft.items.WireCuttersItem;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.misc.PageGroup;
 import net.geforcemods.securitycraft.renderers.DisplayCaseItemRenderer;
@@ -2121,7 +2122,7 @@ public class SCContent {
 	@HasManualPage
 	public static final RegistryObject<Item> UNIVERSAL_OWNER_CHANGER = ITEMS.register("universal_owner_changer", () -> new UniversalOwnerChangerItem(itemProp(SecurityCraft.TECHNICAL_TAB).stacksTo(1).defaultDurability(48)));
 	@HasManualPage
-	public static final RegistryObject<Item> WIRE_CUTTERS = ITEMS.register("wire_cutters", () -> new Item(itemProp(SecurityCraft.TECHNICAL_TAB).stacksTo(1).defaultDurability(476)));
+	public static final RegistryObject<Item> WIRE_CUTTERS = ITEMS.register("wire_cutters", () -> new WireCuttersItem(itemProp(SecurityCraft.TECHNICAL_TAB).stacksTo(1).defaultDurability(476)));
 
 	//modules
 	@HasManualPage