+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadTrapDoorBlock.java
@@ -3,7 +3,7 @@ package net.geforcemods.securitycraft.blocks;
 import java.util.Random;
 
 import net.geforcemods.securitycraft.SCContent;
-import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IPasscodeConvertible;
 import net.geforcemods.securitycraft.api.IPasscodeProtected;
 import net.geforcemods.securitycraft.blockentities.KeypadTrapdoorBlockEntity;
@@ -97,25 +97,39 @@ public class KeypadTrapDoorBlock extends BaseIronTrapDoorBlock {
 
 	public static class Convertible implements IPasscodeConvertible {
 		@Override
-		public boolean isValidStateForConversion(BlockState state) {
+		public boolean isUnprotectedBlock(BlockState state) {
 			return state.is(SCContent.REINFORCED_IRON_TRAPDOOR.get());
 		}
 
 		@Override
-		public boolean convert(PlayerEntity player, World level, BlockPos pos) {
+		public boolean isProtectedBlock(BlockState state) {
+			return state.is(SCContent.KEYPAD_TRAPDOOR.get());
+		}
+
+		@Override
+		public boolean protect(PlayerEntity player, World level, BlockPos pos) {
+			return convert(level, pos, SCContent.KEYPAD_TRAPDOOR.get());
+		}
+
+		@Override
+		public boolean unprotect(PlayerEntity player, World level, BlockPos pos) {
+			return convert(level, pos, SCContent.REINFORCED_IRON_TRAPDOOR.get());
+		}
+
+		private boolean convert(World level, BlockPos pos, Block convertedBlock) {
 			BlockState state = level.getBlockState(pos);
 			Direction facing = state.getValue(FACING);
-			boolean open = state.getValue(OPEN);
 			Half half = state.getValue(HALF);
-			boolean powered = state.getValue(POWERED);
 			boolean waterlogged = state.getValue(WATERLOGGED);
-			TileEntity trapdoor = level.getBlockEntity(pos);
-			CompoundNBT tag = trapdoor.save(new CompoundNBT());
+			TileEntity be = level.getBlockEntity(pos);
+			CompoundNBT tag = be.save(new CompoundNBT());
 
-			level.setBlockAndUpdate(pos, SCContent.KEYPAD_TRAPDOOR.get().defaultBlockState().setValue(FACING, facing).setValue(OPEN, open).setValue(HALF, half).setValue(POWERED, powered).setValue(WATERLOGGED, waterlogged));
-			trapdoor = level.getBlockEntity(pos);
-			trapdoor.load(null, tag);
-			((IOwnable) trapdoor).setOwner(player.getUUID().toString(), player.getName().getString());
+			if (be instanceof IModuleInventory)
+				((IModuleInventory) be).dropAllModules();
+
+			tag = be.save(new CompoundNBT());
+			level.setBlockAndUpdate(pos, convertedBlock.defaultBlockState().setValue(FACING, facing).setValue(OPEN, false).setValue(HALF, half).setValue(POWERED, false).setValue(WATERLOGGED, waterlogged));
+			level.getBlockEntity(pos).load(null, tag);
 			return true;
 		}
 	}